fix CPUTimes type change problem on other packages.

+++ process/process_linux.go
@@ -552,8 +552,8 @@ func (p *Process) fillFromStat() (string, int32, *cpu.CPUTimesStat, int64, int32
 
 	cpuTimes := &cpu.CPUTimesStat{
 		CPU:    "cpu",
-		User:   float32(utime * (1000 / ClockTicks)),
-		System: float32(stime * (1000 / ClockTicks)),
+		User:   float64(utime * (1000 / ClockTicks)),
+		System: float64(stime * (1000 / ClockTicks)),
 	}
 
 	bootTime, _ := host.BootTime()
